Fixing cols stats generation
