+++ hudi-common/src/main/java/org/apache/hudi/common/model/HoodieWriteStat.java
@@ -20,13 +20,11 @@ package org.apache.hudi.common.model;
 
 import org.apache.hudi.common.table.read.HoodieReadStats;
 import org.apache.hudi.common.util.JsonUtils;
-import org.apache.hudi.common.util.Option;
 import org.apache.hudi.storage.StoragePath;
 
 import javax.annotation.Nullable;
 
 import java.io.Serializable;
-import java.util.HashMap;
 import java.util.Map;
 
 /**
@@ -116,8 +114,6 @@ public class HoodieWriteStat extends HoodieReadStats {
   @Nullable
   private RuntimeStats runtimeStats;
 
-  private Option<Map<String, HoodieColumnRangeMetadata<Comparable>>> recordsStats = Option.empty();
-
   public HoodieWriteStat() {
     // called by jackson json lib
   }
@@ -184,24 +180,6 @@ public class HoodieWriteStat extends HoodieReadStats {
 
   public String getPath() {
     return path;
-  }
-
-  public void putRecordsStats(Map<String, HoodieColumnRangeMetadata<Comparable>> stats) {
-    if (!recordsStats.isPresent()) {
-      recordsStats = Option.of(stats);
-    } else {
-      // in case there are multiple log blocks for one write process.
-      recordsStats = Option.of(mergeRecordsStats(recordsStats.get(), stats));
-    }
-  }
-
-  // keep for serialization efficiency
-  public void setRecordsStats(Map<String, HoodieColumnRangeMetadata<Comparable>> stats) {
-    recordsStats = Option.of(stats);
-  }
-
-  public Option<Map<String, HoodieColumnRangeMetadata<Comparable>>> getColumnStats() {
-    return recordsStats;
   }
 
   @Nullable
@@ -365,19 +343,5 @@ public class HoodieWriteStat extends HoodieReadStats {
     public void setTotalCreateTime(long totalCreateTime) {
       this.totalCreateTime = totalCreateTime;
     }
-  }
-
-  private static Map<String, HoodieColumnRangeMetadata<Comparable>> mergeRecordsStats(
-      Map<String, HoodieColumnRangeMetadata<Comparable>> stats1,
-      Map<String, HoodieColumnRangeMetadata<Comparable>> stats2) {
-    Map<String, HoodieColumnRangeMetadata<Comparable>> mergedStats = new HashMap<>(stats1);
-    for (Map.Entry<String, HoodieColumnRangeMetadata<Comparable>> entry : stats2.entrySet()) {
-      final String colName = entry.getKey();
-      final HoodieColumnRangeMetadata<Comparable> metadata = mergedStats.containsKey(colName)
-          ? HoodieColumnRangeMetadata.merge(mergedStats.get(colName), entry.getValue())
-          : entry.getValue();
-      mergedStats.put(colName, metadata);
-    }
-    return mergedStats;
   }
 }
